Moving generation of tooltip to separate file.
