+++ packages/ckeditor5-html-embed/docs/_snippets/features/html-embed.html
@@ -1,5 +1,3 @@
-<button type="button" id="show-me-action">Show me!</button>
-
 <div id="snippet-html-embed"></div>
 
 <button type="button" id="preview-data-action">Preview editor data</button>
